Refactored the PDF components as a universal component.
