+++ src/components/pdf/annotation-box.css
