+++ src/components/pdf/annotation-box.tsx
