+++ src/components/pdf/dictionary-entry.tsx
