+++ src/components/pdf/rect-btns.tsx
